Refactor: Extract ErgogenBranding component
